Project divided in packages, refactoring done correctly
