+++ TrackingApp/app/src/main/java/com/example/trackingapp/Fragments/ConnectionCodeFragment.java
@@ -1,4 +1,4 @@
-package com.example.trackingapp;
+package com.example.trackingapp.Fragments;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -7,6 +7,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
+
+import com.example.trackingapp.R;
 
 public class ConnectionCodeFragment extends androidx.fragment.app.Fragment {
