+++ TrackingApp/app/src/main/java/com/example/trackingapp/Fragments/DatePickerFragment.java
@@ -1,11 +1,10 @@
-package com.example.trackingapp;
+package com.example.trackingapp.Fragments;
 
 import android.app.DatePickerDialog;
 import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.util.Log;
 import android.widget.DatePicker;
 
 import java.util.Calendar;
