+++ TrackingApp/app/src/main/java/com/example/trackingapp/Fragments/FragSettings.java
@@ -1,4 +1,4 @@
-package com.example.trackingapp;
+package com.example.trackingapp.Fragments;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -9,9 +9,9 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.EditText;
 import android.widget.TextView;
-import android.widget.Toast;
+
+import com.example.trackingapp.R;
 
 public class FragSettings extends Fragment {
     //private static final String TAG = "FragSettings";
