+++ TrackingApp/app/src/main/java/com/example/trackingapp/Util/NotificationReceiver.java
@@ -1,11 +1,11 @@
-package com.example.trackingapp;
+package com.example.trackingapp.Util;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.widget.Toast;
 
-import static com.example.trackingapp.Constants.IS_WORKING;
+import static com.example.trackingapp.Util.Constants.IS_WORKING;
 
 public class NotificationReceiver extends BroadcastReceiver {
