+++ TrackingApp/app/src/main/java/com/example/trackingapp/Util/Userinformation.java
@@ -1,4 +1,4 @@
-package com.example.trackingapp;
+package com.example.trackingapp.Util;
 
 import java.util.Map;
