+++ TrackingApp/app/src/main/java/com/example/trackingapp/Util/WriteData.java
@@ -1,12 +1,9 @@
-package com.example.trackingapp;
+package com.example.trackingapp.Util;
 
-import android.app.Dialog;
 import android.app.ProgressDialog;
 import android.content.ContentResolver;
 import android.content.Context;
 import android.net.Uri;
-import android.os.Bundle;
-import android.os.Handler;
 import android.webkit.MimeTypeMap;
 import android.widget.Toast;
 
@@ -15,7 +12,6 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.auth.UserProfileChangeRequest;
 import com.google.firebase.firestore.FirebaseFirestore;
