+++ TrackingApp/app/src/main/res/layout/fragment_settings.xml
@@ -5,7 +5,7 @@
     android:id="@+id/containerSettings"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FragSettings">
+    tools:context=".Fragments.FragSettings">
 
     <TextView
         android:id="@+id/lblAccountSettings"
