+++ TrackingApp/app/src/main/res/xml/pref_headers.xml
@@ -3,7 +3,7 @@
     <!-- These settings headers are only used on tablets. -->
 
     <header
-        android:fragment="com.example.trackingapp.AccountSettings"
+        android:fragment="com.example.trackingapp.Fragments.AccountSettings"
         android:icon="@drawable/ic_info_black_24dp"
         android:title="@string/pref_header_general" />
